// src/app/constants/enums/mj.enum.ts
export enum MjIdEnum {
  TaiJiJianFa = 1,
  QianShuHanMei = 2,
  DuGuJiuJian = 3,
  WanLiuJian = 4,
  LiuMaiShenJian = 5,
  TianWaiFeiXian = 6,
  TianXiaYouXue = 7,
  WuJianZhiJian = 8,
  LuoHanGunFa = 9,
  FengMoZhangFa = 10,
  LaoShiGunFa = 11,
  DaGouBangFa = 12,
  DaMoZhangFa = 13,
  TanZhiShenTong = 14,
  TaiJiQuan = 15,
  NianHuaZhi = 16,
  AnRanXiaoHunZhang = 17,
  FengWuJiuTian = 18,
  XiangLongShiBaZhang = 19,
  QingRenJian = 20,
  KongQueLing = 21,
  XiaoLiFeiDao = 22,
  LingXiYiZhi = 23,
  WangQingTianShu = 24,
  JiuYangZhenJing = 25,
  JiuYinZhenJing = 26,
  YiJinJing = 27,
  ZuoYouHuBo = 28,
  HuanShiBiShen = 29,
  BeiMingShenGong = 30,
}